better than last commit
Intent:
Any illegal write latches the addr and value that was attempted to
write.
If the write occured in the 0xFF__ region, it is considered a syscall,
and its offset into that region indicates what syscall it was, where the
value written is the argument to the syscall.
Otherwise it is just considered a segfault